+++ kite/src/main/kotlin/com/tang/kite/executor/Executor.kt
@@ -1,5 +1,6 @@
 package com.tang.kite.executor
 
+import com.tang.kite.sql.BatchSqlStatement
 import com.tang.kite.sql.SqlStatement
 import java.sql.Connection
 
@@ -15,6 +16,10 @@ interface Executor : AutoCloseable {
     fun <T> query(statement: SqlStatement, type: Class<T>): List<T>
 
     fun update(statement: SqlStatement, parameter: Any): Int
+
+    fun update(batchSqlStatement: BatchSqlStatement, parameters: List<Any>): Int
+
+    fun update(statements: List<SqlStatement>, parameters: List<Any>): Int
 
     fun commit()