+++ kite/src/main/kotlin/com/tang/kite/sql/BatchSqlStatement.kt
@@ -1,0 +1,12 @@
+package com.tang.kite.sql
+
+/**
+ * @author Tang
+ */
+class BatchSqlStatement(
+
+    val sql: String,
+
+    val parameters: MutableList<MutableList<Any?>>
+
+)