+++ kite/src/main/kotlin/com/tang/kite/utils/Statements.kt
@@ -1,6 +1,12 @@
 package com.tang.kite.utils
 
 import java.sql.PreparedStatement
+import java.sql.Timestamp
+import java.sql.Types
+import java.time.Instant
+import java.time.LocalDate
+import java.time.LocalDateTime
+import java.time.LocalTime
 
 /**
  * Statement utils
@@ -26,6 +32,10 @@ object Statements {
             is Short -> statement.setShort(index, parameter)
             is Float -> statement.setFloat(index, parameter)
             is Double -> statement.setDouble(index, parameter)
+            is LocalDate -> statement.setObject(index, Timestamp.valueOf(parameter.atStartOfDay()), Types.DATE)
+            is LocalTime -> statement.setObject(index, parameter, Types.TIME)
+            is LocalDateTime -> statement.setObject(index, Timestamp.valueOf(parameter), Types.TIMESTAMP)
+            is Instant -> statement.setObject(index, Timestamp.from(parameter))
             is ByteArray -> statement.setBytes(index, parameter)
             else -> statement.setObject(index, parameter)
         }