+++ kite/src/test/kotlin/com/tang/kite/session/entity/Account.kt
@@ -4,7 +4,7 @@ import com.tang.kite.annotation.Column
 import com.tang.kite.annotation.id.Id
 import com.tang.kite.annotation.id.IdType
 import java.math.BigDecimal
-import java.sql.Date
+import java.time.LocalDateTime
 
 /**
  * @author Tang
@@ -15,9 +15,9 @@ class Account (
     var id: Long? = null,
     var username: String? = null,
     var password: String? = null,
-    var createTime: Date? = null,
+    var createTime: LocalDateTime? = null,
     @Column("update_time")
-    var updateTime: Date? = null,
+    var updateTime: LocalDateTime? = null,
     var balance: BigDecimal? = null,
     @Column(ignore = true)
     var beIgnore: String? = null